initial commit
I didn't want to go 3 rounds with the proper version of ansible for some
random syntax, so I kept it as simple as possible with a little bit of
bash scripting.

I know there are things that are incorrect. This is a baseline copy/paste
from the docs to get us started.